+++ fs_utils.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const path = require("path");
+const CWD = process.cwd();
 
 function createFolder(path) {
   if (!fs.existsSync(path)) {
@@ -202,6 +203,66 @@ function getSharedParts(firmId, handle) {
   return sharedPartsPresent;
 }
 
+// List all files of a specific type (recursive search)
+// Return an array with their full paths
+function listExistingFiles(typeCheck = "liquid") {
+  const baseDirectory = fs.readdirSync(CWD);
+  const basePath = path.resolve(CWD);
+  const array = recursiveInspectDirectory({
+    basePath: basePath,
+    collection: baseDirectory,
+    pathsArray: undefined,
+    typeCheck: typeCheck,
+  });
+  return array;
+}
+
+// Recursive option for fs.watch is not available in every OS (e.g. Linux)
+function recursiveInspectDirectory({
+  basePath,
+  collection,
+  pathsArray = [],
+  typeCheck = "liquid",
+}) {
+  collection.forEach((filePath) => {
+    let fullPath = path.resolve(basePath, filePath);
+    let fileStats = fs.statSync(fullPath, () => {});
+
+    if (fileStats.isDirectory()) {
+      let directory = fs.readdirSync(fullPath);
+      recursiveInspectDirectory({
+        basePath: fullPath,
+        collection: directory,
+        pathsArray: pathsArray,
+        typeCheck: typeCheck,
+      });
+    }
+    let fileType = fullPath.split(".")[fullPath.split.length - 1];
+    if (fileType === typeCheck) {
+      pathsArray.push(fullPath);
+    }
+  });
+  return pathsArray;
+}
+
+// Return {type, handle} of a template
+// type: reconciliationText | sharedPart
+function identifyTypeAndHandle(filePath) {
+  let index;
+  const pathParts = path.resolve(filePath).split(path.sep);
+  const sharedPartCheck = (element) => element === "shared_parts";
+  const reconciliationCheck = (element) => element === "reconciliation_texts";
+  index = pathParts.findIndex(sharedPartCheck);
+  if (index !== -1) {
+    return { type: "sharedPart", handle: pathParts[index + 1] };
+  }
+  index = pathParts.findIndex(reconciliationCheck);
+  if (index !== -1) {
+    return { type: "reconciliationText", handle: pathParts[index + 1] };
+  }
+  return false;
+}
+
 module.exports = {
   readConfig,
   writeConfig,
@@ -214,4 +275,6 @@ module.exports = {
   getTemplatePaths,
   findHandleByID,
   getSharedParts,
+  listExistingFiles,
+  identifyTypeAndHandle,
 };
